feat: 재시도 조건 수정 및 예외 추가 처리
- 요청한 인덱스가 범위 내 인덱스 여부인지 판단 후, 아니라면 INVALID_LOCATION_INDEX 예외 발생
- 삭제하고자 하는 인덱스가 마지막 인덱스일 때에는 갱신된 rows 수 검사하는 로직 적용 X
- 인덱스 범위 검사 후, 인덱스를 찾을 수 없는 경우에는 동시성 문제이므로, CONCURRENT_MODIFICATION_EXCEPTION 예외를 발생시키도록 설정

+++ src/main/java/com/genius/herewe/business/location/facade/DefaultLocationFacade.java
@@ -4,10 +4,10 @@ import static com.genius.herewe.core.global.exception.ErrorCode.*;
 
 import java.util.List;
 
+import org.hibernate.StaleStateException;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.genius.herewe.business.location.LocationRequest;
@@ -90,23 +90,31 @@ public class DefaultLocationFacade implements LocationFacade {
 				return;
 			} catch (Exception e) {
 				retryCount++;
-				if (shouldRetry(e, retryCount)) {
+				if (shouldRetry(e) && retryCount < MAX_RETRIES) {
 					logRetryAttempt(retryCount, momentId, locationIndex, e);
 					applyBackoff(retryCount);
-					continue;
+				} else {
+					throw translateException(e);
 				}
-				throw translateException(e);
 			}
 		}
-		log.error("모든 재시도 실패");
 	}
 
-	private boolean shouldRetry(Exception e, int retryCount) {
-		// 최대 재시도 횟수 미만이고 재시도 가능한 예외인 경우
-		return retryCount < MAX_RETRIES &&
-			(e instanceof DataAccessException ||
-				e instanceof OptimisticLockException ||
-				(e.getCause() != null && e.getCause() instanceof DataAccessException));
+	private boolean shouldRetry(Exception e) {
+		boolean isDbException = e instanceof OptimisticLockException
+			|| e instanceof DataAccessException
+			|| e instanceof StaleStateException;
+
+		boolean isRetryException = (e instanceof BusinessException) &&
+			(((BusinessException)e).getErrorCode() == CONCURRENT_MODIFICATION_EXCEPTION);
+
+		boolean hasRetryCause = e.getCause() != null && (
+			e.getCause() instanceof DataAccessException ||
+				(e.getCause().getMessage() != null &&
+					e.getCause().getMessage().contains("Deadlock"))
+		);
+
+		return isDbException || isRetryException || hasRetryCause;
 	}
 
 	private void applyBackoff(int retryCount) {
@@ -133,20 +141,27 @@ public class DefaultLocationFacade implements LocationFacade {
 		}
 	}
 
-	@Transactional(propagation = Propagation.REQUIRES_NEW)
+	@Transactional
 	private void executeDeletePlace(Long userId, Long momentId, int locationIndex) {
 		momentMemberService.findByJoinInfo(userId, momentId)
 			.orElseThrow(() -> new BusinessException(MOMENT_PARTICIPATION_NOT_FOUND));
 
 		Moment moment = momentService.findByIdWithOptimisticLock(momentId);
+		int lastIndex = locationService.findLastIndexForMoment(momentId);
+		if (1 > locationIndex || locationIndex > lastIndex) {
+			throw new BusinessException(INVALID_LOCATION_INDEX);
+		}
 
 		Location targetLocation = locationService.findByIndex(momentId, locationIndex)
-			.orElseThrow(() -> new BusinessException(LOCATION_NOT_FOUND));
+			.orElseThrow(() -> new BusinessException(CONCURRENT_MODIFICATION_EXCEPTION));
 		locationService.delete(targetLocation);
 
-		int updatedRows = locationService.bulkDecreaseIndexes(momentId, locationIndex, moment.getVersion());
-		if (updatedRows == 0) {
-			throw new BusinessException(CONCURRENT_MODIFICATION_EXCEPTION);
+		boolean isLastIndex = locationIndex == lastIndex;
+		if (!isLastIndex) {
+			int updatedRows = locationService.bulkDecreaseIndexes(momentId, locationIndex, moment.getVersion());
+			if (updatedRows == 0) {
+				throw new BusinessException(CONCURRENT_MODIFICATION_EXCEPTION);
+			}
 		}
 
 		moment.updateLastModifiedTime();
